Feature05 - UnionMemberShip 	Modified Files: 		Union.java: 			- Changed argument type in constructor from int to double 		Admin.java: 			- Added loadUnion() and dumpUnion() in the interface and implemented it 			- Added addUnion() and deleteUnion() in the interface

+++ SourceCode/Union.java
@@ -24,7 +24,7 @@ class Union{
 		this.members = new ArrayList<>();
 		this.donations = new ArrayList<>();
 	}
-	public Union(String name, int membershipfees){
+	public Union(String name, double membershipfees){
 		this();
 		this.name = name;
 		this.membershipfees = membershipfees;
